webapp/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pandas>=1.5.0
scikit-learn>=1.2.0
numpy>=1.23.0
scipy>=1.9.0
Pillow>=10.0.0
pytesseract>=0.3.10
werkzeug>=2.3.0


